requirements.txt
altair==6.1.0
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
blinker==1.9.0
cachetools==7.0.6
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
cryptography==46.0.7
distro==1.9.0
gitdb==4.0.12
GitPython==3.1.46
google-ai-generativelanguage==0.6.15
google-api-core==2.30.3
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-genai==1.73.1
googleapis-common-protos==1.74.0
groq==1.2.0
grpcio==1.80.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.12
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
mysql-connector-python==9.6.0
narwhals==2.20.0
numpy==2.4.4
packaging==26.1
pandas==3.0.2
pillow==12.2.0
proto-plus==1.27.2
protobuf==5.29.6
pyarrow==24.0.0
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.3
pydantic_core==2.46.3
pydeck==0.9.2
pydub==0.25.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
referencing==0.37.0
requests==2.33.1
rpds-py==0.30.0
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
SpeechRecognition==3.16.0
streamlit==1.56.0
streamlit-audiorec @ git+https://github.com/stefanrmmr/streamlit-audio-recorder.git@777d18114130137d492c0378a86631fff1ff1be5
streamlit_mic_recorder==0.0.8
tenacity==9.1.4
toml==0.10.2
tornado==6.5.5
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
watchdog==6.0.0
websockets==16.0
